Fixed logger error and commented some code
